requirements.txt
aiofiles==24.1.0
dill==0.3.8
faiss==1.8.0
fastapi==0.115.0
Jinja2==3.1.4
loguru==0.7.2
matplotlib==3.9.2
numpy==2.1.1
openai==1.51.0
pandas==2.2.3
peft==0.12.0
pydantic==2.9.2
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
ruamel.base==1.0.0
sentence_transformers==3.0.1
starlette==0.39.2
torch==2.4.0
tqdm==4.66.4
transformers==4.44.2
trl==0.11.0
uvicorn==0.31.0
wandb==0.18.3
